obs-transitions: Fix integer conversion warning

+++ plugins/obs-transitions/transition-stinger.c
@@ -63,7 +63,7 @@ static void stinger_update(void *data, obs_data_t *settings)
 	else
 		s->transition_point_ns = (uint64_t)(point * 1000000LL);
 
-	s->monitoring_type = obs_data_get_int(settings,"audio_monitoring");
+	s->monitoring_type = (int)obs_data_get_int(settings,"audio_monitoring");
 	obs_source_set_monitoring_type(s->media_source, s->monitoring_type);
 
 	s->fade_style = (enum fade_style)obs_data_get_int(settings,
